Add table for displaying data file status (public/private)
This commit involves creating a table component that displays the status (public or private) of data files. The table is designed to allow users to easily view the status of all data files in a single view. This table will help users to easily identify which data files are public and which are private.